aptos-agent/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
aptos-sdk==0.10.0
attrs==24.3.0
behave==1.2.6
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
ecdsa==0.18.0
filelock==3.16.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
identify==2.6.4
idna==3.10
iniconfig==2.0.0
instructor==1.7.2
Jinja2==3.1.5
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
nodeenv==1.9.1
numpy==2.2.1
oauthlib==3.2.2
openai==1.58.1
packaging==24.2
parse==1.20.2
parse_type==0.6.4
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
propcache==0.2.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyNaCl==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
python-graphql-client==0.4.3
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
swarm @ git+https://github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
tenacity==9.0.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
types-six==1.17.0.20241205
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.28.0
websockets==14.1
yarl==1.18.3
